chore(control-panel): upload station and upgrader separately
