+++ tests/integration/src/control_panel_tests.rs
@@ -1,7 +1,7 @@
-use crate::setup::{
-    get_canister_wasm, setup_new_env, setup_new_env_with_config, SetupConfig, WALLET_ADMIN_USER,
+use crate::setup::{setup_new_env, setup_new_env_with_config, SetupConfig, WALLET_ADMIN_USER};
+use crate::utils::{
+    controller_test_id, upload_canister_modules, user_test_id, NNS_ROOT_CANISTER_ID,
 };
-use crate::utils::{controller_test_id, user_test_id, NNS_ROOT_CANISTER_ID};
 use crate::TestEnv;
 use candid::Principal;
 use control_panel_api::{
@@ -478,22 +478,7 @@ fn no_upload_canister_modules() {
         .unwrap()
         .contains("Canister config not initialized"));
 
-    // upload canister modules
-    let upgrader_wasm = get_canister_wasm("upgrader").to_vec();
-    let station_wasm = get_canister_wasm("station").to_vec();
-    let upload_canister_modules_args = UploadCanisterModulesInput {
-        station_wasm_module: station_wasm.to_owned(),
-        upgrader_wasm_module: upgrader_wasm.to_owned(),
-    };
-    let res: (ApiResult<()>,) = update_candid_as(
-        &env,
-        canister_ids.control_panel,
-        controller,
-        "upload_canister_modules",
-        (upload_canister_modules_args.clone(),),
-    )
-    .unwrap();
-    res.0.unwrap();
+    upload_canister_modules(&env, canister_ids.control_panel, controller);
 
     // deploying user station succeeds after uploading canister modules
     let deploy_station_args = DeployStationInput {
@@ -524,22 +509,12 @@ fn upload_canister_modules_authorization() {
         ..
     } = setup_new_env();
 
-    let upgrader_wasm = get_canister_wasm("upgrader").to_vec();
-    let station_wasm = get_canister_wasm("station").to_vec();
-    let upload_canister_modules_args = UploadCanisterModulesInput {
-        station_wasm_module: station_wasm.to_owned(),
-        upgrader_wasm_module: upgrader_wasm.to_owned(),
-    };
-    let res: (ApiResult<()>,) = update_candid_as(
-        &env,
-        canister_ids.control_panel,
-        controller,
-        "upload_canister_modules",
-        (upload_canister_modules_args.clone(),),
-    )
-    .unwrap();
-    res.0.unwrap();
+    upload_canister_modules(&env, canister_ids.control_panel, controller);
 
+    let upload_canister_modules_args = UploadCanisterModulesInput {
+        station_wasm_module: None,
+        upgrader_wasm_module: None,
+    };
     let res: (ApiResult<()>,) = update_candid_as(
         &env,
         canister_ids.control_panel,
